requirements.txt
flask
requests
openai
python-dotenv